llm/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
backoff==2.2.1
    # via coinbase-advanced-py
boto3==1.35.68
    # via -r requirements.in
botocore==1.35.68
    # via
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
coinbase-advanced-py==1.8.2
    # via -r requirements.in
cryptography==44.0.2
    # via coinbase-advanced-py
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.9.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
openai==1.68.2
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pyjwt==2.10.1
    # via coinbase-advanced-py
python-dateutil==2.9.0.post0
    # via botocore
python-ulid==3.0.0
    # via -r requirements.in
requests==2.32.3
    # via coinbase-advanced-py
s3transfer==0.10.4
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
structlog==24.4.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via
    #   botocore
    #   requests
websockets==13.1
    # via coinbase-advanced-py
